1) Script now creates a histogram of all unfiltered (so valid for sending to customer) messages before capping function and limit take place. 2) Couple of structural changes
